# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
...
  handlers.EnvelopeError:
    properties:
      error:
        example: invalid id
        type: string
      success:
        example: false
        type: boolean
    type: object
  handlers.EnvelopeOK:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  handlers.createVMRequest:
    properties:
      enable_gpu:
        type: boolean
      generate_key:
        type: boolean
      hostname:
        type: string
      os_template:
        type: string
      public_tunnel:
        type: boolean
      required_tags:
        description: |-
          RequiredTags is the host-aggregate filter as a CSV string
          (e.g. "fast-cpu,nvme"). Operator-defined free-form labels —
          no enum validation. Empty = no constraint.
        type: string
      ssh_key_id:
        type: integer
      ssh_privkey:
        type: string
      ssh_pubkey:
        type: string
      subdomain:
        type: string
      tier:
        type: string
      tunnel_port:
        type: integer
    type: object
...
  handlers.deleteTunnelResponse:
    properties:
      message:
        example: tunnel deleted
        type: string
    type: object
...
  handlers.privateKeyResponse:
    properties:
      key_name:
        type: string
      private_key:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Self-hosted VM provisioning portal for Proxmox VE. The API is
    wrapped in a uniform `{success, data, error}` envelope on every
    response — see EnvelopeOK / EnvelopeError schemas below.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nimbus HTTP API
  version: "1.0"
paths:
...
  /vms:
    get:
      description: |-
        Includes legacy rows that pre-date owner tracking (owner_id
        NULL). Live status is fetched from Proxmox at request time.
        Cluster-wide views are admin-only and live under /cluster/vms.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/db.VM'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: List the caller's VMs
      tags:
      - vms
    post:
      consumes:
      - application/json
      description: |-
        Streams `application/x-ndjson` once headers flush — one
        `{type:"progress", step, label}` line per phase, then a
        terminal `{type:"result", data: ...}` (success) or
        `{type:"error", code, message}` (failure). Total wall time
        30-60s for the provision, plus 1-3min if Gopher bootstrap
        waits on dpkg locks during cloud-init contention. Validation
        failures pre-stream return the standard {success,error}
        envelope with the appropriate 4xx status.
      parameters:
      - description: VM provisioning request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createVMRequest'
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: NDJSON stream of progress events terminated by a result or
            error line
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Provision a new VM (NDJSON event stream)
      tags:
      - vms
  /vms/{id}:
    delete:
      description: |-
        Cross-user and legacy (owner_id NULL) rows return 404 by
        design — existence is not disclosed across ownership
        boundaries. Admin-side destruction goes through
        DELETE /cluster/vms/{id} instead.
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Destroy one of the caller's VMs
      tags:
      - vms
    get:
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/db.VM'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Get one of the caller's VMs by id
      tags:
      - vms
  /vms/{id}/{op}:
    post:
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      - description: Lifecycle op
        enum:
        - start
        - shutdown
        - stop
        - reboot
        in: path
        name: op
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Run a power op on the caller's VM
      tags:
      - vms
  /vms/{id}/private-key:
    get:
      description: |-
        404 when no private key was deposited for this VM. Owner-gated:
        non-owners get 404 (existence not disclosed).
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/handlers.privateKeyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Download the private SSH key for one of the caller's VMs
      tags:
      - vms
  /vms/{id}/tunnels:
    get:
      description: |-
        Empty array for VMs without a Gopher machine record and for
        tunnel-disabled deployments. Owner-gated.
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/tunnel.Tunnel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: List per-port tunnels attached to one of the caller's VMs
      tags:
      - vms
    post:
      consumes:
      - application/json
      description: Mirrors Gopher's tunnel-creation body shape.
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      - description: Per-port tunnel spec
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createTunnelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/tunnel.Tunnel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Attach a per-port tunnel to one of the caller's VMs
      tags:
      - vms
  /vms/{id}/tunnels/{tunnelId}:
    delete:
      parameters:
      - description: VM DB id
        in: path
        name: id
        required: true
        type: integer
      - description: Gopher tunnel id
        in: path
        name: tunnelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/handlers.deleteTunnelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Detach a per-port tunnel from one of the caller's VMs
      tags:
      - vms
...
securityDefinitions:
  cookieAuth:
    description: |-
      Session cookie set by /api/auth/login. Required by all routes
      outside the public auth + setup surface.
    in: cookie
    name: nimbus_sid
    type: apiKey
  workerAuth:
    description: |-
      Bearer token used by the GX10 GPU worker daemon. Format:
      `Bearer <hex>`. Mint via the Settings → GPU page.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"